+++ utils.js
@@ -1,0 +1,7 @@
+// Generic error handler used by all endpoints.
+function handleError(res, reason, message, code) {
+  console.log(`ERROR: ${reason}`);
+  res.status(code || 500).json({ error: message });
+}
+
+exports.handleError = handleError;
